@@ -7,9 +7,9 @@
     "lib": "lib"
   },
   "scripts": {
-    "test": "npm run generate-d-ts",
-    "build": "npm run generate-d-ts",
-    "generate-d-ts": "rm lib/index.d.ts lib/ioTsBrands.d.ts lib/ioTsTypes.d.ts && tsc"
+    "run-tests": "npx ts-mocha tests/**/*.ts",
+    "test": "tsc && npm run run-tests",
+    "build": "tsc"
   },
   "repository": {
     "type": "git",
@@ -23,13 +23,18 @@
   "homepage": "https://github.com/imin-ltd/speck#readme",
   "husky": {
     "hooks": {
-      "pre-commit": "npm run build && git add lib/index.d.ts lib/ioTsBrands.d.ts"
+      "pre-commit": "npm run build && git add lib"
     }
   },
   "devDependencies": {
+    "@types/chai": "^4.2.16",
     "@types/lodash": "^4.14.149",
+    "@types/mocha": "^8.2.2",
+    "chai": "^4.3.4",
     "husky": "^4.2.1",
-    "typescript": "^3.7.5"
+    "mocha": "^8.3.2",
+    "ts-mocha": "^8.0.0",
+    "typescript": "^4.2.4"
   },
   "engines": {
     "node": "10.16.0",
