@@ -1,1 +1,6 @@
-export const ioTsJsDate: import("io-ts").Type<Date, Date, unknown>;
+/**
+ * Custom io-ts types. See how to create new io-ts types here:
+ * https://github.com/gcanti/io-ts/blob/master/index.md#the-idea.
+ */
+import * as t from 'io-ts';
+export declare const ioTsJsDate: t.Type<Date, Date, unknown>;
